@@ -5,12 +5,13 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.os.Parcelable
-import android.support.v4.app.Fragment
-import android.support.v7.widget.GridLayoutManager
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.GridLayoutManager
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.chad.library.adapter.base.BaseQuickAdapter
@@ -37,7 +38,7 @@ import java.util.*
  * Created by D22434 on 2017/11/29.
  */
 
-class HomePageFragment : Fragment(), HomePageContract.View, BaseQuickAdapter.RequestLoadMoreListener {
+class HomePageFragment : androidx.fragment.app.Fragment(), HomePageContract.View, BaseQuickAdapter.RequestLoadMoreListener {
 
     private var mPresenter: HomePagePresenter = HomePagePresenter()
     private var mAdapter: HomeAdapter = HomeAdapter(null)
@@ -64,9 +65,9 @@ class HomePageFragment : Fragment(), HomePageContract.View, BaseQuickAdapter.Req
         return inflater.inflate(R.layout.fragment_homepage, container, false)
     }
 
-    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view!!, savedInstanceState)
-        order = arguments.getString("order")
+        order = arguments?.getString("order") ?: ""
         mPresenter.attachView(this)
         initView()
         initAdapter()
@@ -75,7 +76,7 @@ class HomePageFragment : Fragment(), HomePageContract.View, BaseQuickAdapter.Req
 
     private fun initView() {
         recyclerView.isNestedScrollingEnabled = false
-        recyclerView.layoutManager = GridLayoutManager(activity, 3)
+        recyclerView.layoutManager = androidx.recyclerview.widget.GridLayoutManager(activity, 3)
     }
 
     private fun loadData() {
@@ -144,7 +145,7 @@ class HomePageFragment : Fragment(), HomePageContract.View, BaseQuickAdapter.Req
                 intent.putExtra("cover", banners[position].lcover)
                 intent.putExtra("name", banners[position].name)
                 intent.putExtra("type", "album")
-                context.startActivity(intent)
+                context?.startActivity(intent)
             }
         })
         bannerView.banner.setIndicatorVisible(false)
@@ -158,20 +159,22 @@ class HomePageFragment : Fragment(), HomePageContract.View, BaseQuickAdapter.Req
     }
 
     override fun showError(error: String) {
-        mAdapter.emptyView.findViewById<LoadingView>(R.id.loading_view).setLoadingText(context.getText(R.string.load_error))
+        mAdapter.emptyView.findViewById<LoadingView>(R.id.loading_view).setLoadingText(context?.getText(R.string.load_error))
         mAdapter.loadMoreFail()
     }
 
     inner class BannerViewHolder : MZViewHolder<Banner> {
         private var mView: View? = null
         override fun onBind(p0: Context?, p1: Int, banner: Banner?) {
-            GlideApp.with(context)
-                    .load(banner?.lcover)
-                    .placeholder(R.drawable.ic_default_preview)
-                    .centerCrop()
-                    .diskCacheStrategy(DiskCacheStrategy.NONE)
-                    .transition(DrawableTransitionOptions.withCrossFade())
-                    .into(mView?.findViewById(R.id.imageView))
+            context?.let {
+                GlideApp.with(it)
+                        .load(banner?.lcover)
+                        .placeholder(R.drawable.ic_default_preview)
+                        .centerCrop()
+                        .diskCacheStrategy(DiskCacheStrategy.NONE)
+                        .transition(DrawableTransitionOptions.withCrossFade())
+                        .into(mView?.findViewById<ImageView>(R.id.imageView)!!)
+            }
         }
 
         @SuppressLint("InflateParams")
