@@ -2,8 +2,8 @@ package com.dragon.wallpaperapp.ui.fragment
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.LinearLayoutManager
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -23,7 +23,7 @@ import kotlinx.android.synthetic.main.fragment_movie.*
  * Created by D22434 on 2017/11/29.
  */
 
-class MovieFragment : Fragment(), MovieContract.View {
+class MovieFragment : androidx.fragment.app.Fragment(), MovieContract.View {
 
     var mPresenter = MoviePresenter()
     var mAdapter = SectionAdapter(R.layout.item_movie, R.layout.item_header, null)
@@ -32,7 +32,7 @@ class MovieFragment : Fragment(), MovieContract.View {
         return inflater.inflate(R.layout.fragment_movie, container, false)
     }
 
-    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view!!, savedInstanceState)
         init()
         mPresenter.attachView(this)
@@ -41,7 +41,7 @@ class MovieFragment : Fragment(), MovieContract.View {
 
     private fun init() {
         recyclerView.isNestedScrollingEnabled = false
-        recyclerView.layoutManager = LinearLayoutManager(activity)
+        recyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(activity)
         mAdapter.bindToRecyclerView(recyclerView)
         mAdapter.onItemClickListener = BaseQuickAdapter.OnItemClickListener { adapter, _, position ->
             // 在这里可以做任何自己想要的处理
@@ -50,7 +50,7 @@ class MovieFragment : Fragment(), MovieContract.View {
                 val intent = Intent(context, MovieDetailActivity::class.java)
                 intent.putExtra(MovieDetailActivity.WEB_URL, data.flimData?.url)
                 intent.putExtra(MovieDetailActivity.NAME, data.flimData?.title)
-                context.startActivity(intent)
+                context?.startActivity(intent)
             }
         }
     }
